Restructured js into folders Updated with a placeholder image for lack of stock photos from api formatted frontend into danish
