+++ src/main/resources/static/index.html
@@ -10,7 +10,7 @@
             integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
             crossorigin="anonymous"></script>
     <link rel="stylesheet" href="css/main.css">
-    <script defer src="js/main.js"></script>
+    <script type="module" defer src="js/main.js"></script>
 </head>
 
 <body>
@@ -21,9 +21,10 @@
             <h1 class="mb-4">Tilbud på fødevarer</h1>
 
             <div class="mb-3">
-                <label for="postalCode" class="form-label">Indtast postnummer</label>
+                <p>Find gode tilbud på fødevarer i dit nabolag, og få idéer til opskrifter</p>
+                <label for="postalCode" class="form-label">Indtast dit postnummer</label>
                 <input type="text" id="postalCode" class="form-control" placeholder="f.eks. 2300">
-                <button type="button" class="btn btn-primary mt-2" id="btn-get-offers">Find Tilbud</button>
+                <button type="button" class="btn btn-primary mt-2" id="btn-get-offers">Søg tilbud</button>
             </div>
 
             <div id="offers-container" class="row g-4">
@@ -34,19 +35,18 @@
         <!-- Right side: Recipe Request Section -->
         <div class="col-md-4">
             <div class="fixed-sidebar">
-                <h2>Recipe Assistant</h2>
-                <p>Request a recipe based on a simple topic below:</p>
+                <h1>Opskriftsassistent</h1>
+                <p>Få en opskrift baseret på en madvare, du har valgt</p>
 
                 <div class="card">
                     <div class="card-body">
-                        <h5 class="card-title">Get Recipe</h5>
-                        <label for="about" class="form-label">Recipe Topic</label>
-                        <input type="text" id="about" class="form-control mb-3" placeholder="Enter a recipe topic">
-                        <button type="button" class="btn btn-primary" id="btn-get-recipe">Get Recipe</button>
+                        <h5 class="card-title">Indtast en madvare</h5>
+                        <input type="text" id="about" class="form-control mb-3" placeholder="f.eks. Banan">
+                        <button type="button" class="btn btn-primary" id="btn-get-recipe">Få opskriftsidé</button>
 
                         <!-- Spinner for loading indication -->
                         <div id="spinner" class="spinner-border text-success mt-3" role="status" style="display:none;">
-                            <span class="visually-hidden">Loading...</span>
+                            <span class="visually-hidden">Indlæser...</span>
                         </div>
 
                         <!-- Recipe result displayed here -->
