+++ src/main/resources/static/js/ui/renderRecipe.js
@@ -1,0 +1,3 @@
+export const renderRecipe = (data, resultElement) => {
+    resultElement.innerHTML = data.answer.replace(/\n/g, "<br>") || "No recipe found for the given topic.";
+};
